fix: Addressed some comments in the PR
