--- client/api/teams/types.ts
@@ -1,9 +1,0 @@
-interface TeamResponse {
-  id: number;
-  group_id: number;
-  name: string;
-  created_at: Date;
-  updated_at: Date;
-}
-
-export type { TeamResponse };
