+++ client/components/Forms/Form.module.css
@@ -7,7 +7,6 @@
 
 .inputContainer {
   display: flex;
-  gap: 8px;
   justify-content: space-between;
   align-items: baseline;
 }
