+++ client/pages/Seasons/types.ts
@@ -1,8 +1,6 @@
 interface SeasonType {
   id: number;
   name: string;
-  // start_date: Date;
-  // end_date: Date;
   start_date: string;
   end_date: string;
   is_active: boolean;
