+++ server/controllers/season.controller.js
@@ -206,7 +206,6 @@ async function isNameUniqueWithinLeague(name, leagueId) {
       name: name,
     },
   });
-  console.log(league);
 
   return league === null;
 }
